Created helper for complete internship and moved methods from model there.
