+++ config/locales/complete_internships/de.yml
@@ -8,3 +8,4 @@ de:
     country: Land
     start_date: Von
     end_date: Bis
+    internship_state: Praktikumsstatus
