+++ config/locales/complete_internships/en.yml
@@ -8,3 +8,4 @@ en:
     country: Country
     start_date: Start
     end_date: End
+    internship_state: Internship State
